backend/requirements.txt
# Backend Dependencies

# FastAPI Framework
fastapi==0.124.4
uvicorn[standard]==0.38.0
python-multipart==0.0.21

# Database
sqlalchemy==2.0.45
alembic==1.17.2
psycopg[binary]==3.3.2

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
bcrypt==4.2.1

# Validation
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0

# Utilities
redis==7.1.0
redis[asyncio]
celery[redis]==5.6.0
flower==2.0.1

# AWS S3
boto3==1.35.76

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==11.0.0

# Text Processing
# tiktoken==0.8.0  # Commented out - requires Rust compiler on Windows, will use character-based chunking instead
langdetect==1.0.9

# WebSocket
websockets==12.0
